git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/trunk@3135 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ magick/animate.c
@@ -2280,7 +2280,8 @@ MagickExport Image *XAnimateImages(Display *display,
                     pixel->border_color.pixel;
                   magick_windows[i]->attributes.colormap=map_info->colormap;
                   (void) XChangeWindowAttributes(display,magick_windows[i]->id,
-                    magick_windows[i]->mask,&magick_windows[i]->attributes);
+                    (unsigned long) magick_windows[i]->mask,
+                    &magick_windows[i]->attributes);
                 }
                 if (windows->backdrop.id != (Window) NULL)
                   (void) XInstallColormap(display,map_info->colormap);
